Move getMetadataDataTypes to separate file (helper, for now)
