+++ src/scripts/modules/transformations/react/components/mapping/InputMappingRowSnowflakeEditor.jsx
@@ -13,6 +13,7 @@ import {PanelWithDetails} from '@keboola/indigo-ui';
 import MetadataStore from '../../../../components/stores/MetadataStore';
 import createStoreMixin from '../../../../../react/mixins/createStoreMixin';
 import { SnowflakeDataTypesMapping } from '../../../Constants';
+import { getMetadataDataTypes } from './InputMappingRowSnowflakeEditorHelper';
 
 export default React.createClass({
   propTypes: {
@@ -169,61 +170,12 @@ export default React.createClass({
     return null;
   },
 
-  getMetadataDataTypes(columnMetadata) {
-    return columnMetadata.map((metadata, colname) => {
-      let datatypeLength = metadata.filter((entry) => {
-        return entry.get('key') === 'KBC.datatype.length';
-      });
-      if (datatypeLength.count() > 0) {
-        datatypeLength = datatypeLength.get(0);
-      }
-      let datatypeNullable = metadata.filter((entry) => {
-        return entry.get('key') === 'KBC.datatype.nullable';
-      });
-      if (datatypeNullable.count() > 0) {
-        datatypeNullable = datatypeNullable.get(0);
-      }
-      let basetype = metadata.filter((entry) => {
-        return entry.get('key') === 'KBC.datatype.basetype';
-      });
-
-      if (basetype.count() === 0) {
-        return null;
-      } else {
-        basetype = basetype.get(0);
-      }
-      let datatypeName, length = null;
-
-      let datatype = SnowflakeDataTypesMapping.map((mappedDatatype) => {
-        if (mappedDatatype.get('basetype') === basetype.get('value')) {
-          datatypeName = mappedDatatype.get('name');
-          return mappedDatatype;
-        }
-      });
-      const mapType = datatype.get(datatypeName);
-      if (mapType) {
-        length = mapType.get('size') ? datatypeLength.get('value') : null;
-        if (mapType.has('maxLength') && length > mapType.get('maxLength')) {
-          length = mapType.get('maxLength');
-        }
-      }
-      return fromJS({
-        column: colname,
-        type: datatypeName,
-        length: length,
-        convertEmptyValuesToNull: isNaN(datatypeNullable.get('value'))
-          ? datatypeNullable.get('value')
-          : !!parseInt(datatypeNullable.get('value'), 10)
-      });
-    });
-  },
-
   getDefaultDatatypes() {
     if (this.state.hasMetadataDatatypes) {
       const metadataSet = this.state.tableColumnMetadata.filter((metadata, colname) => {
         return this._getFilteredColumns().indexOf(colname) > -1;
       });
-      return this.getMetadataDataTypes(metadataSet);
+      return getMetadataDataTypes(metadataSet);
     } else {
       return fromJS(this._getFilteredColumns()).reduce((memo, column) => {
         return memo.set(column, fromJS({
@@ -238,7 +190,7 @@ export default React.createClass({
 
   getInitialDatatypes(sourceTable) {
     if (MetadataStore.tableHasMetadataDatatypes(sourceTable)) {
-      return this.getMetadataDataTypes(MetadataStore.getTableColumnsMetadata(sourceTable));
+      return getMetadataDataTypes(MetadataStore.getTableColumnsMetadata(sourceTable));
     } else {
       return this.props.tables.find((table) => {
         return table.get('id') === sourceTable;
